requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
python=3.8.18
numpy=1.24.3
pytorch=2.1.2
pytorch-cuda=11.8
scipy=1.10.1
tensorboard=2.12.1